@@ -1,5 +1,5 @@
 import React, { PureComponent } from 'react'
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 
 import {
 	AdminContainer,
@@ -30,6 +30,19 @@ import {
 	Tooltip,
 } from 'components/bits'
 
+const testPrint = () =>{
+	const stuff = (
+	<>
+		<Routes>
+			<Route path='/' element={<LandingContainer />} />
+			<Route path='/search-public-collections' element={<SearchPublicCollectionsContainer />}/>
+			<Route element={<Error error='404' message={`You've wandered too far`} />}/>
+		</Routes>
+	</>
+	)
+	console.log(stuff) //eslint-disable-line no-console
+	return stuff
+}
 class Root extends PureComponent {
 
 	render() {
@@ -39,7 +52,7 @@ class Root extends PureComponent {
 			loading,
 			modal,
 		} = this.props.viewstate
-
+		console.log(user) //eslint-disable-line no-console
 		// TODO: route has to be touched mirroring with backend
 		return (
 			<Router>
@@ -47,108 +60,96 @@ class Root extends PureComponent {
 					<>
 						<HeaderContainer />
 						<MenuContainer />
-						<Switch>
+						<Routes>
 
-							<Route exact path='/' >
-								<CollectionsContainer />
-							</Route>
-
-							<Route exact path='/search-public-collections' >
-								<SearchPublicCollectionsContainer />
-							</Route>
+							<Route path='/' element={<CollectionsContainer />}/>
+							<Route path='/search-public-collections' element={<SearchPublicCollectionsContainer />}/>
 
 							{
 								user.roles === 0 &&
-								<Route path='/admin'>
-									<AdminContainer />
-								</Route>
+								<Route path='/admin' element={<AdminContainer />}/>
 							}
 
 							{
 								user.roles < 3 &&
-							<Route path='/lab-assistant'>
-								<LabAssistantContainer />
-							</Route>
+							<Route path='/lab-assistant' element={<LabAssistantContainer />}/>
 							}
 
-							<Route path='/collections'>
-								<CollectionsContainer />
-							</Route>
+							<Route path='/collections' element={<CollectionsContainer />}/>
 
 							{
 								user.roles < 3 &&
-								<Route path='/manage-resource'>
-									<ManageResourceContainer />
+								<Route path='/manage-resource' element={<ManageResourceContainer />}/>
+							}
+							{
+								user.roles < 3 &&
+								<Route path='/lab-assistant-manager' element={<LabAssistantManagerContainer />}>
+									<Route path=':professorId' element={<LabAssistantManagerContainer />}>
+										<Route path=':collectionId' element={<LabAssistantManagerContainer />}/>
+									</Route>
 								</Route>
 							}
 							{
 								user.roles < 3 &&
-								<Route path='/lab-assistant-manager/:professorId/:collectionId?'>
-									<LabAssistantManagerContainer />
+								<Route path='/manager' element={<ManagerContainer />}>
+									<Route path =':id' element={<ManagerContainer />}/>
 								</Route>
 							}
 							{
 								user.roles < 3 &&
-								<Route path='/manager/:id?'>
-									<ManagerContainer />
-								</Route>
-							}
-							{
-								user.roles < 3 &&
-								<Route path='/public-manager/:id?'>
-									<PublicManagerContainer />
-								</Route>
+								<Route path='/public-manager' element={<PublicManagerContainer />}>
+									<Route path=':id' element={<PublicManagerContainer />}/>
+									</Route>
 							}
 
-							<Route path='/player/:id/:clip?'>
-								<PlayerContainer />
+							<Route path='/player' element={<PlayerContainer />}>
+								<Route path=':id' element={<PlayerContainer />}>
+									<Route path=':clip' element={<PlayerContainer />}/>
+								</Route>
 							</Route>
 							{
 								user.roles < 3 &&
-								<Route path='/videoeditor/:id'>
-									<VideoEditorContainer />
+								<Route path='/videoeditor' element={<VideoEditorContainer />}>
+									<Route path=':id' element={<VideoEditorContainer />}/>
 								</Route>
 							}
 
-							<Route path='/subtileeditor/:id'>
-								<SubtitlesEditorContainer />
+							<Route path='/subtileeditor'>
+								<Route path=':id' element={<SubtitlesEditorContainer />}/>
 							</Route>
 
-							<Route path='/clipeditor/:id'>
-								<ClipEditorContainer />
+							<Route path='/clipeditor' element={<ClipEditorContainer />}>
+								<Route path=':id' element={<ClipEditorContainer />}/>
 							</Route>
 
 							{
 								user.roles < 3 &&
-								<Route path='/clipeditor/:id'>
-									<ClipEditorContainer />
-								</Route>
+								<Route path='/clipeditor' element={<ClipEditorContainer />}>
+								<Route path=':id' element={<ClipEditorContainer />}/>
+							</Route>
 							}
-							<Route path='/feedback'>
-								<FeedbackContainer />
-							</Route>
+							<Route path='/feedback' element={<FeedbackContainer />}/>
 
-							<Route>
-								<Error error='404' message={`You've wandered too far`} />
-							</Route>
-						</Switch>
+							<Route element={<Error error='404' message={`You've wandered too far`} />}/>
+						</Routes>
 					</>
 					:
 					(
-						<>
-							<Switch>
-								<Route exact path='/'>
-									<LandingContainer />
-								</Route>
-								<Route exact path='/search-public-collections' >
-									{/* <MenuContainer /> */}
-									<SearchPublicCollectionsContainer />
-								</Route>
-								<Route>
-									<Error error='404' message={`You've wandered too far`} />
-								</Route>
-							</Switch>
-						</>
+						// <>
+						// 	<Routes>
+						// 		<Route exact path='/'>
+						// 			<LandingContainer />
+						// 		</Route>
+						// 		<Route exact path='/search-public-collections' >
+						// 			<MenuContainer />
+						// 			<SearchPublicCollectionsContainer />
+						// 		</Route>
+						// 		<Route>
+						// 			<Error error='404' message={`You've wandered too far`} />
+						// 		</Route>
+						// 	</Routes>
+						// </>
+						testPrint()
 					)
 				}
 
@@ -159,5 +160,4 @@ class Root extends PureComponent {
 		)
 	}
 }
-
 export default Root
