@@ -1,7 +1,7 @@
 /* eslint-disable function-paren-newline */
 
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 
 import * as serviceWorker from './serviceWorker'
 
@@ -9,14 +9,13 @@ import { Provider } from 'react-redux'
 import store from 'services/store'
 
 import { RootContainer } from 'containers'
-
-ReactDOM.render(
+const cont = 	document.getElementById(`root`)
+const root = ReactDOM.createRoot(cont)
+root.render(
 	<Provider store={store}>
 		<RootContainer />
 	</Provider>,
-	document.getElementById(`root`),
 )
-
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: http://bit.ly/CRA-PWA
