@@ -1,9 +1,10 @@
 import React, { useState, useEffect, useRef } from 'react'
-import { Prompt } from 'react-router'
+// import { Prompt } from 'react-router'
 import Style, { Timeline, EventList, Icon } from './styles'
 import { Rnd } from 'react-rnd'
 import { SubtitleEditorSideMenu, SubtitlesCard, SubtitlesLayer, SwitchToggle } from 'components/bits'
 import * as Subtitle from 'subtitle'
+import {parse} from 'subtitle'
 
 import { VideoContainer, SkipLayer } from 'components'
 import { convertToSeconds } from '../../common/timeConversion'
@@ -78,7 +79,6 @@ const SubtitleEditor = props => {
 			eventsArray.sort((a, b) => a.layer > b.layer ? 1 : -1)
 			largestLayer = eventsArray[eventsArray.length-1].layer
 		}
-
 		// Find the largest layer number
 		const initialLayers = []
 
@@ -246,8 +246,7 @@ const SubtitleEditor = props => {
 					document.getElementById(`subStart${index}`).style.border=`2px solid red`
 					needCheck = false
 				}
-			}
-			else if (side === `end`) {
+			} else if (side === `end`) {
 				if(sub.end.match(/^\d{2}:\d{2}\.\d{2}/) !== null || sub.end.match(/^\d{1}:\d{2}:\d{2}\.\d{2}/) !== null || type === `onBlur`) {
 					sub.end = convertToSeconds(sub.end, videoLength)
 					document.getElementById(`subEnd${index}`).style.border=null
@@ -459,10 +458,12 @@ const SubtitleEditor = props => {
 		setBlock(true)
 	}
 	const handleAddSubLayerFromFile = (url) => {
+		debugger
 		try{
 			const reader = new FileReader()
 			reader.onload = (e) => {
-				const temp = Subtitle.parse(e.target.result)
+				console.log(e) //eslint-disable-line
+				const temp = parse(e.target.result)
 				// console.log(Subtitle.parse(e.target.result))
 				for (let i = 0; i < temp.length; i++){
 					temp[i].start = temp[i].start /1000
@@ -790,8 +791,8 @@ const SubtitleEditor = props => {
 								}
 								className={`setSubModalVisible`}
 								onClick={ () => {
-								  openSubModal(isReady, setIsReady, `create`, ``, handleAddSubLayer, handleAddSubLayerFromFile)
-							   }}>
+									openSubModal(isReady, setIsReady, `create`, ``, handleAddSubLayer, handleAddSubLayerFromFile)
+								}}>
 								<p id={`editIcon`} style={{ fontWeight:700 }}>Add Subtitle Track +</p>
 							</div>
 						</div>
@@ -926,10 +927,10 @@ const SubtitleEditor = props => {
 				</>
 			</EventList>
 			<>
-				<Prompt
+				{/* <Prompt
 					when={blockLeave}
 					message='If you leave you will lose all your changes. Are you sure to leave without saving?'
-				/>
+				/> */}
 			</>
 		</Style>
 	)
