@@ -16,19 +16,20 @@ const SubtitlesContainer = props => {
 
 		}else
 			setSubtitlesArray([])
+		for(let i = 0; i<subtitlesArray.length; i++){
+			const element = subtitlesArray[i]
+			const start = element.start
+			const end = element.end
+			if(currentTime >= start && currentTime <= end){
+				handleShowSubtitle(element.text)
+				break
+			}else if (currentTime > end || currentTime < start)
+				handleShowSubtitle(``)
+		}
 			// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [duration, subtitles])
 
-	for(let i = 0; i<subtitlesArray.length; i++){
-		const element = subtitlesArray[i]
-		const start = element.start
-		const end = element.end
-		if(currentTime >= start && currentTime <= end){
-			handleShowSubtitle(element.text)
-			break
-		}else if (currentTime > end || currentTime < start)
-			handleShowSubtitle(``)
-	}
+
 
 	const eventClassArray = subtitles
 	const viewstate = {
